Move createLogger
