+++ src/cli/util.js
@@ -20,6 +20,7 @@ const optionsNormalizer = require("../main/options-normalizer");
 const thirdParty = require("../common/third-party");
 const arrayify = require("../utils/arrayify");
 const isTTY = require("../utils/is-tty");
+const logger = require("../utils/logger");
 
 const OPTION_USAGE_THRESHOLD = 25;
 const CHOICE_USAGE_MARGIN = 3;
@@ -764,54 +765,6 @@ function pick(object, keys) {
       );
 }
 
-function createLogger(logLevel) {
-  return {
-    warn: createLogFunc("warn", "yellow"),
-    error: createLogFunc("error", "red"),
-    debug: createLogFunc("debug", "blue"),
-    log: createLogFunc("log")
-  };
-
-  function createLogFunc(loggerName, color) {
-    if (!shouldLog(loggerName)) {
-      return () => {};
-    }
-
-    const prefix = color ? `[${chalk[color](loggerName)}] ` : "";
-    return function(message, opts) {
-      opts = Object.assign({ newline: true }, opts);
-      const stream = process[loggerName === "log" ? "stdout" : "stderr"];
-      stream.write(message.replace(/^/gm, prefix) + (opts.newline ? "\n" : ""));
-    };
-  }
-
-  function shouldLog(loggerName) {
-    switch (logLevel) {
-      case "silent":
-        return false;
-      default:
-        return true;
-      case "debug":
-        if (loggerName === "debug") {
-          return true;
-        }
-      // fall through
-      case "log":
-        if (loggerName === "log") {
-          return true;
-        }
-      // fall through
-      case "warn":
-        if (loggerName === "warn") {
-          return true;
-        }
-      // fall through
-      case "error":
-        return loggerName === "error";
-    }
-  }
-}
-
 function normalizeDetailedOption(name, option) {
   return Object.assign({ category: coreOptions.CATEGORY_OTHER }, option, {
     choices:
@@ -917,7 +870,7 @@ function createContext(args) {
   updateContextArgv(context);
   normalizeContextArgv(context, ["loglevel", "plugin", "plugin-search-dir"]);
 
-  context.logger = createLogger(context.argv["loglevel"]);
+  context.logger = logger.createLogger(context.argv["loglevel"]);
 
   updateContextArgv(
     context,
