+++ src/utils/logger.js
@@ -1,0 +1,55 @@
+"use strict";
+
+const chalk = require("chalk");
+
+function createLogger(logLevel) {
+  return {
+    warn: createLogFunc("warn", "yellow"),
+    error: createLogFunc("error", "red"),
+    debug: createLogFunc("debug", "blue"),
+    log: createLogFunc("log")
+  };
+
+  function createLogFunc(loggerName, color) {
+    if (!shouldLog(loggerName)) {
+      return () => {};
+    }
+
+    const prefix = color ? `[${chalk[color](loggerName)}] ` : "";
+    return function(message, opts) {
+      opts = Object.assign({ newline: true }, opts);
+      const stream = process[loggerName === "log" ? "stdout" : "stderr"];
+      stream.write(message.replace(/^/gm, prefix) + (opts.newline ? "\n" : ""));
+    };
+  }
+
+  function shouldLog(loggerName) {
+    switch (logLevel) {
+      case "silent":
+        return false;
+      default:
+        return true;
+      case "debug":
+        if (loggerName === "debug") {
+          return true;
+        }
+      // fall through
+      case "log":
+        if (loggerName === "log") {
+          return true;
+        }
+      // fall through
+      case "warn":
+        if (loggerName === "warn") {
+          return true;
+        }
+      // fall through
+      case "error":
+        return loggerName === "error";
+    }
+  }
+}
+
+module.exports = {
+  createLogger
+};
